requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
googleapis-common-protos==1.70.0
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.6
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.0
livekit==1.0.13
livekit-agents==1.2.9
livekit-api==1.0.5
livekit-blingfire==1.0.0
livekit-plugins-cartesia==1.0.20
livekit-plugins-noise-cancellation==0.2.2
livekit-plugins-openai==1.0.20
livekit-plugins-rime==1.2.9
livekit-plugins-silero==1.0.11
livekit-plugins-tavus==1.2.9
livekit-plugins-turn-detector==1.0.11
livekit-protocol==1.0.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
numpy==2.0.2
onnxruntime==1.19.2
openai==1.107.3
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pillow==11.3.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.56.1
types-protobuf==6.30.2.20250914
typing-inspection==0.4.1
typing_extensions==4.15.0
Unidecode==1.3.8
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0

